+++ cypress/integration/createContent.js
@@ -37,5 +37,15 @@ if (Cypress.env('API') !== 'guillotina') {
       );
       // cy.get('.view-wrapper a').click();
     });
+    it('Create Image', function() {
+      cy.createContent('Image', 'my-image', 'My Image');
+      cy.visit('/my-image');
+      cy.get('.documentFirstHeading').should('have.text', 'My Image');
+      cy.get('.view-wrapper img')
+        .should('have.attr', 'src')
+        .and('include', '/my-image/@@images/');
+      cy.get('.view-wrapper img').should('have.attr', 'alt', 'My Image');
+      // cy.get('.view-wrapper a').click();
+    });
   });
 }